fix/refactor(amplitude): Add pad function, trim so amp not affected by taper
Previously there was a hacky bit of code in scan.py which was supposed to ensure the
raw_waveforms cut by read_event_waveform_data() were sufficiently long that the
taper would not interfere with the noise or signal amplitude measurements. This code
was dodgy, and did not do the job.

This new function explicitly handles the calculation of the padding required such that
the earliest possible noise and latest possible signal windows are guaranteed not to be
affected by the taper applied during response removal / filtering.

Furthermore, the raw_waveforms are now trimmed to this time window. Previously, even
if the start was early enough and the end was late enough the taper could affect the
measurement windows if the pre- and post-pad were sufficiently lopsided.