fix(cut_waveforms): Allow pre/post_cut to be float or None
We work with pre / post_cut as either None or float within scan.py, but the
check in cut_waveforms.py was `if pre_cut != 0.0`. This was unnecessarily
specific, and let None's through to be added to event.otime --> breaks.

The comparison has been changed to `if pre_cut` which catches both 0.0 and None.