refactor(amplitude): Change how overlapping P/S windows are distinguished
Previously a fairly obscure process was used to set new end and start times for
the P and S signal windows where they overlapped. Now they are simply divided at
the midway time.

The previous approach intended to preserve the P signal window where the phase
arrivals were very close in time in comparison to the window lengths, but this
might lead to undesirable behaviour in edge cases / where the user makes rogue
parameter choices. Overall not worth it.

NOTE: as might be expected this makes a small difference to most P and S signal
amplitude measurements. A thorough check through the changes to the dike example
suggest these changes are all sensible.