style/docs(data.py): Make it explicit we are using `nearest_sample=True`
See the ObsPy docs on :func:`obspy.core.trace.Trace.trim` for details; this has
consequences for the behaviour when reading or trimming data between a start and
end time which do not fall directly on timestamps corresponding to samples of the
waveform data. This is the default behaviour, so previously it was implicit.

Also a minor change to the definitions of the `data.starttime` and
`data.endtime` attributes, which are not strictly linked to the actual
timestamps of the data (which may contain gaps, or be recorded at a different
sampling rate to `onset.sampling_rate`); instead these are the start and
end times of the period of data that was *requested*.