feat(data): Add function to handle off-sample timestamped data
Occasionally data is recorded "off-sample"; i.e. the timestamps are a non-
integer number of samples after midnight. An example would be 100Hz data
with timestamps of 00:00:003 00:00:013 ...

This commit adds a new function to correct this data. This is necessary as
the migration machinery within QM takes the waveform data as an array, so
cannot handle inconsistent timestamps between traces from different
stations. Two options are provided: to shift the data by simply shifting
the metadata - this is what was implicitly being done previously. However,
this leads to a sub-sample timing offset. Alternatively, the data can be
interpolated (using the lanczos interpolation scheme provided by obspy)
to shift the timing of the samples while maintaining the timing of the
data.

Currently this function is applied when data is read in from the archive
and added to data.waveforms (used for onset calculation and migration);
importantly data.raw_waveforms is not altered.

Plus a tiny bit of variable renaming in util.decimate (no consequence for
its function).