fix/refactor(resample): Add padding when upsampling
Previously when data was upsampled during resampling (e.g. if going from
raw data sampling rate of 40 Hz to an onset function sampling rate of
50 Hz), no check was made that the start and endtimes of the data read
from the archive exactly matched the start and endtimes requested (which
are calculated assuming the data being read is at the onset function
sampling rate, or a multiple if it). In many cases this would not be
the case, leading to this data being needlessly discarded.

Building from the shift_to_sample() function added in the previous commit,
it is now possible to straightforwardly calculate the amount of additional
padding required when upsampling - this is added as a constant value pad,
though is not important as it will be within the pre-pad period of the
data read from the archive, and will be tapered. All data is then
strictly trimmed between the requested start and endtime, now using the
`nearest_sample=False` option.

Within stalta, this resampled data is then double-checked with the new
check_start_and_endtimes option in data.check_availability to ensure
the data being fed into the migration meets all the implicit and
explicit requirements.