feat(stalta): Add function to set onset=0 in tapering windows.
The data read from the archive is padded to allow for the pre- and post-pad
needed for the particular onset function algorithm (a function of the sta
and lta windows for STALTA), plus an additional padding to account for the
region of data that will be tapered when applying filters etc.

This region of the onset function is implicitly disregarded during migration,
because the padding windows have been calculated such that this time window
is never migrated into the grid. However, it is still used in calculating
an appropriate pick threshold in the GaussianPicker. The erroneously high
onset function values at the start of the onset function (where the taper
leads to a high values of the STA/LTA) can therefore interfere with this.

Here a function is added to set the onset function in these time windows
to 0, if the timespan of data being migrated is provided. This will
allow more appropriate pick thresholds to be calculated in gaussianpicker.