feat/fix(gaussianpicker): Trim tapered windows of onset function
The functionality added in the previous commit is used to set the value
of the onset function in the taper-pad regions to 0, allowing a more
representative pick threshold to be calculated.

This inevitably leads to changes in the pick thresholds, but should not
significantly affect the pick times, uncertainties or SNR. There will be
minor effects due to fitting to a greater proportion of the data, in some
cases.