fix(gaussianpicker): Determine threshold from onset > 1
Changing the filter to only consider the data > 1 means that data during the
taper windows for calculating the onset function is ignored, as well as data
during the pick windows.

This test can be improved when the logging / clipping etc. of the onset
functions is moved from onset() to migrate().