fix(interpolate): Add pad after interpolating.
The new option to shift the timing by interpolation of the original data
had only been tested with positive sub-sample time shifts, and with an
additional data pad being cut for local magnitude calculation. Where
shifting by a negative time, or running without any additional data
padding, the strict constraint that interpolate will *not*
extrapolate outside the timespan of the input trace leads to crashes.

For negative time shifts, the `new_starttime` is now shifted by one sample
so it is within the timespan of the input data. A single-sample constant-
value pre-pad is added subsequently so the data is the correct length.
This data is contained within the taper, so will not have any influence
on the results.

For positive time shifts, a single-sample constant-value pad is likewise
added at the end of the interpolated trace.