chore(benchmarks): Update benchmarks after changes made in #115
As described in the commits within #115, changes to the pick thresholds
calculated from the onset function data outside the pick windows will
cause (in general) more picks to be made, and will cause minor changes
to the timing, amplitude and uncertainty of picks that were made
previously, due to this change in the threshold affecting how much
data is fitted to when making the pick.

These changes have been carefully reviewed and all appear as expected.