feat/refactor(gaussianpicker): Add MAD pick threshold option
A second method has been added to determine the pick threshold in
GaussianPicker. The new option uses the Median Absolute Deviation of the
onset function outside the pick window(s) to determine the pick threshold;
as the median plus a multiple of the MAD.

In adding this option, the syntax for specifying the pick threshold has
changed; the user must now specify both the `threshold_method` as well as the
threshold relevant to that option (`percentile_pick_threshold` or
`mad_pick_threshold`). The previous `pick_threshold` parameter has been
deprecated with an instructive error message.

The `pick_threshold` attribute has been removed from event.picks, because it is not
used, and adding multiple options therefore seems more faff than it's worth.