refactor(gaussianpicker): Remove 88th percentile threshold adjustment
Previously an arbitrary secondary pick threshold was calculated from the
88th percentile of the onset function data within the relevant pick window.
The higher of this `signal_threshold` and the user-facing pick threshold
would then be set as the threshold above which a gaussian would be fit to
the onset data.

Having looked more closely at how often this was actually the case, and its
effect when used, the behind-the-scenes nature of this approach seems clearly
not justified by any clear improvement in pick performance. It has therefore
been removed.

This will result in occasional differences in the reported pick threshold and
fitting (pick) parameters where all other settings are otherwise equal. But
hopefully the improvement in transparency will justify this.