fix(amplitudes): Set max P_amp window length to `signal_window`
Previously the end of the P window would always be the same as the start of
the S_amp window. This ignores the user-specified `signal_window` parameter,
meaning they can't control the P signal window length (even if that is the
phase that is specified for magnitudes calculation). This changes the
overlap handling so that P_end = S_start if the gap is smaller than
`signal_window`, but otherwise is set by the signal window.