refactor(amplitudes): Return avg amp in signal windows, filter gain.
Additionally measure the average amplitude in the P and S signal windows,
using the same method as for the noise window. Comparison of these measures
is likely to prove useful for filtering out regional and teleseismic
events which are falsely located in the grid.

As part of this change, noise_amps will now be returned *not* corrected for
any filter gain, and the filter gains corrected for in the P and S amplitude
measurements (both p2p and average) will be recorded. This allows for the
noise amplitude to be corrected by the same factor for comparison to the P
or S amplitude. Previously it was automatically corrected for the filter gain
appropriate for the S psp measurement. The new arrangement gives greater
flexibility to the user.

In accordance with this, the appropriate filter gain is now corrected for
when calculating magnitudes.

The conversion from metres -> millimetres has been moved into the
_peak_to_trough_amplitude() function.

NOTE: the changes in this commit obviously change the format of the .amps file,
leading to the benchmark tests to fail. This will be remedied once the changes
in this PR are approved.