feat/refactor(magnitudes): Add option for custom filtering for mag r2
Add a user-facing (though marked as experimental) option to use a custom
filter for which amplitude observations to consider for mag r-squared
statistic calculation. Observations with low SNR are rejected for magnitude
calculation (with a user-specified threshold). However, the absence of
high amplitude observations where they are predicted to be seen based on
the event location and magnitude are of strong importance to the mag r2
calculation. This custom filter compares the noise level to the predicted
amplitude where amplitudes have been discarded due to the SNR filter, and
if the noise amplitude is more than 5 times smaller than the predicted
amplitude, these observations are retained for r-squared calculation.

Also, rename the internal function _amp_r_squared() to _mag_r_squared().
(No external impact).