refactor(get_wa_response): Split get_wa_waveform to get_real / get_wa
The function to get the Woods-Anderson simulated waveform for a Trace was
an all-in-one that included deconvolving the instrument response and
simulating the Woods-Anderson response. This was convenient, but to add the
ability to efficiently generate real-response waveforms (to optionally
output for use in further analysis) this has now been split into two
functions. The first gets the real waveform (by removing the instrument
response), and the second simulates the WA response (after calling the
get_real_waveform function to get the real waveform).

This leads to very small changes in the measured amplitudes, on the order of
1 in 10^-6, due to doing a small additional amount of processing between
removing the instrument response and simulating the WA (= detrending,
essentially), where previously it was done in a single call to
tr.simulate(). There *should* be no difference for `remove_full_response`.

This also now makes it possible to obtain the Woods-Anderson simulated
velocity waveform, if requested. Note that this all attenuation functions
currently included in the QM local_mags module are for displacement
seismograms.