refactor(response removal): Shift response removal to Archive / WaveformData
The response removal parameters were previously provided with the amplitude
measurement parameters to `local_mags.amplitudes`. This doesn't make much sense,
as they are then all passed to functions within `WaveformData` to do the actual
response removal.

They have now been shifted to `Archive` - specified with the response inventory -
and then are likewise added to the `WaveformData` object served up from an
Archive query. They are then readily available for response removal from any
sub-module, not just `local_mags`.

This requires a slight change to the run scripts (updated templates to follow
in the next commit), but no change to the functionality.