fix(Archive): Always assign `water_level` etc. attributes
@hemmelig uncovered a bug which I had missed - `water_level` etc. need to
be set to be able to pass them on to the `WaveformData` object, whether
a `response_inv` has been provided or not. Their assignment has therefore
been moved outside the if, instead just using that to raise the warning
about setting the `water_level`.

All examples now run without any errors.