fix/refactor(staltaonset): Move log to `ilib.migrate()`
The log(onsets) operation is performed as part of the geometric mean stacking
operation; taking the log allows the GM to be calculated as a sum rather than
a product. Previously the log was taken within calculate_onsets(), which
doesn't make much sense. Furthermore, it was taken *before* combining
horizontal components to create a single S-wave onset function.

By moving the log operation to migrate() its purpose is made far more clear,
and this bug is fixed. This will lead to minor changes to results where the
user includes more than one component in calculating (e.g.) the S-wave
onset function.