refactor/fix(stalta.py): Set null values to 1
Null values where the sta/lta cannot be calculated should be set to 1 for
our purposes, not 0, as we are using it for onset detection rather than
classical triggering.

In order to do this, rather than using the obspy `classic_sta_lta` function
a customised version of it has been added to stalta.py.