feat/refactor(stalta.py): Add multiple options for signal transformation
To compute the STA/LTA of a signal, the signal must be transformed such that
it is always non-negative. Previously this has been achieved by taking the
square of the signal; here 3 new options are added.

1) The absolute value of the signal
2) The absolute value of the analytic signal (envelope)
3) The square of the signal envelope (~ equivalent to the energy of the
   signal, and most similar to what has been the default previously).

Which of these options provides the most faithful representation of the
PDF of a phase onset will depend on the characteristics of the phase
arrivals and noise.