refactor/feat(stalta): Make `signal_transform`, `min_onset_value` class att's
The `method` parameter introduced in the previous commit has been replaced by
a class variable `signal_transform` which does the same job (but a lot more
neatly).

In addition, the option to clip onsets at some minimum value has been
implemented - this was theoretically here previously, though written incorrectly.

This gives the user the flexibility to set this clip (equivalent to a minimum
SNR filter) at an appropriate value for their dataset and choice of
`signal_transform`. However it should be noted that it doesn't have a strong
effect as long as it is set at a sensible value. It also serves the function
to prevent errors arising from trying to take the log of 0 (in migrate()).