docs: Add Archive tutorial
Add a short tutorial covering the basics of Archive, including how to
specify the waveform archive structure, instrument response removal, and
the various criteria that are used to determine whether data is
'available' for a given timestep.

This is a re-work of #77 in light of a string of changes to where data
availability is handled in QuakeMigrate.