fix(cut_waveforms): Handling of ResponseNotFoundError, empty traces
The catch for a `util.ResponseNotFoundError` was incorrectly placed at the
top-level call to get_waveforms() for a given event. This meant that if a
single station / trace was not able to be matched to a response in the
provided response inventory, the function would return (and no waveform
data be written). This is now re-raised with a helpful explanation.

Checks are also added that the traces are not empty (particularly after
trimming to the `pre_cut` and `post_cut`, if applicable) before attempting
to remove / simulate the response, and that the Stream is not empty,
before attempting to write it to file.

Plus some minor docstrings etc. tidying.