fix(amplitude): Catch case where amplitude measurement freq = Nyquist
The approximate nature of our frequency estimate for the amplitude
measurement makes it possible for it to be equal to (or minutely
higher than) the Nyquist frequency of the data. If a bandpass filter has
been applied, its gain is not defined above the Nyquist, so a query for
the absolute value of the filter gain would be returned zero, leading
to an overflow and an amplitude measurement of inf.

This case is now caught, returning an info-level message reporting the
approx_freq and phase of the failed measurement, and the trace summary
string.