tests: Add Travis CI tests
Add Travis Continuous Integration - this will allow us to catch breaking
issues before new features/fixes are merged and, in particular, should
help us keep on top of changes related to upstream changes to
dependencies. Both the Iceland icequake and Iceland dike examples are
tested.

This includes a few adjustments to the existing tests and some new ones:

- New import tests to ensure everything has been installed and works
nicely before running the examples
- Test all .event, .amps, and .picks files. Ensure these are correctly
matched by sorting the globbed files
- Switch to using the the pandas.testing function to ensure DataFrames
are equivalent (up to floating-point errors)

This has been extensively tested on a remote fork and everything passes.

Also included is an automatic assessment of the 'code coverage' (i.e.
how many lines of the package are run and validated), which is uploaded
to codecov.io on a successful test!