feat(make_lut): Change lut method name "1dsweep" -> "1dnlloc"
Name change for increased clarity (the sweeping business is not user-facing).
The template script and dike intrusion example are update to reflect this,
and a deprecation handler has been added and tested.

Also a significant batch of docstring improvements / fixes within
create_lut.py, plus the addition of an AttributeError to handle the case
where a user specifies an invalid method for traveltime computation.