feat(notebooks): Set single-threaded for binder compatibility
With a clear message to increase the thread count if the user's system
allows.

Also add a pick plot example to the Rutford example. The same can be done
for the icequake example when the parameter choices have been updated.

Plus a couple of minor formatting tweaks (let's see how thorough stickler
is with notebooks....)