fix(onset): Handle case where no data passes availability check (#126)
This resolves the issue raised in #126 - for a complete description please
read the issue.

This is achieved by adding a new custom `DataAvailabilityException`
exception which is raised where the sum of the availability dict = 0; i.e.
no data available for onset calculation for any phase. Note that it may
be wise to bump this number slightly (e.g. < 3), but just implementing
what's needed for the bug fix for now.

The `ArchiveEmptyException` and `DataGapException` messages have also
been updated to reflect the recent changes in the data reading/processing
pipeline.

And, finally, the dtype of the zeros appended to the availability DataFrame
are now specified as int's. Likely just a cosmetic issue, but still an
improvement to be consistent.