fix(data): Fix potential bug with `read_all_stations=True`
When `read_all_stations=True`, the `station` field in the archive format
string is specified as `*`. Depending on the format string, this can lead
to the string containing `**`, which since Python 3.5 is interpreted by
glob as a call to perform a recursive search -- *if* `recursive=True` --
and otherwise is not passed to glob at all.

This results in the glob call only returning the parent directory which
contains the data (e.g. for the dike intrusion example <blah>/2014/236)
which then throws an error because `obspy.read()` can't read in
directories..

Fixed with a simple "replace("**", "*"). Note this is only applied to
the `read_all_stations=True` case; otherwise the user is relied upon to
not set a format string which can include isolated "**"'s.