feat: Prepare everything for PyPI
Prepare everything required for the registration of QuakeMigrate with
the Python Package Index (PyPI). This will enable users to install the
package using pip as:

`pip install quakemigrate`

The following changes were required:

- Added a MANIFEST.in file that specifies the header files for the C
extension
- Corrected the README.md and docs/index.rst files
- Added Python version requirements. Must be using 3.6, 3.7, or 3.8.
Things do not currently work with 3.9 as there are no wheels available
for some of the specific versions, namely matplotlib 3.2.2. ObsPy is
also yet to be fully available for Python 3.9
- Adjusted setup.py to properly handle the requirement for numpy to be
available during the installation process.
- Added explicit requirement for PyQt5. This is used as the matplotlib
backend, which has been the case for much of the development. We felt it
safer to just add this as a requirement for now, and investigate whether
or not it is strictly required later on.

Work remaining:
- Need to figure out how to provide binary distributions ('wheels')
which contain a precompiled version of the package that should help us
bypass some of the difficulties we face regarding Operating Systems,
compilers, etc etc...