:wrench: Refactor setup.py
Refactor setup.py to be a bit more like a linear recipe, and make some
general improvements/clarity changes.

- Add a full docstring at the top of the file including a short
description, a longer description (taken from the README), and a
description of the license
- Clean up imports
- Make the script 'fail fast' if the incorrect version of Python is
being used
- Check the platform being used for evidence that it is Windows and
create a flag that can be used later to help build the extension
- Make install/extra requirements clear early on in the script
- Change how the Windows export_symbols hack is incorporated. Was
previously fed in within a custom class (inheriting from build_ext), but
this seems to be how ObsPy does things and I received no complaints when
trialling on different operating systems (Linux/macOS principally)
- Removed various functions that were making the script very non-linear
- Rewrote get_extensions to perform all of the jobs of the above
scripts, in a linear fashion. Took advantage of understanding a bit
better the different platform types and their requirements

I also binned off the function that was reading meta data from the top
__init__.py file. The information is still in their, within the
docstring, but I figured there was very little reason to go searching
for other files when it could all be in here (it was all quite static
anyway).

Use setuptools to find the packages within QM, rather than manually
listing them. If any new packages are added, we won't need to remember
to add them here.

This is all preparation for using GitHub Actions to build the various
wheels for QM and upload to PyPI.

No issues with installation. I ran the 2 examples for which we have benchmarks and validated that
everything remained identical.