:wrench: Add Action to build wheels
Add a GitHub Action workflow to build the wheels (hyped up .zip files
that can be used to distribute Python packages with compiled elements
pre-compiled against various Python versions/operating
systems/architectures). This workflow is a product of a tutorial,
influence from ObsPy, and significant testing against a separate
package. For now, I am just including a trial build against each
operating system for a single Python version (3.7).

I have also included a pyproject.toml file, which allows us to pass
arguments and specifications to workflows. Here, I specify that a
pre-requisite for building QM is numpy (v1.21, for ObsPy
compatibility...) as well as specific instructions for macOS operating
systems to preinstall llvm and libomp.

This workflow is activated whenever something is PUSHED to master OR any
branch prefixed with "test_wheels". If a tag is pushed to the GH repo,
then it will also attempt to upload the resulting wheels to Test PyPI.
For this to be successful, we need to set up SECRET tokens for our
TestPyPI/PyPI accounts. In future, we can also configure it to publish
properly to PyPI when a new version is released through GitHub (by
simply uncommenting the final section).

This will only run in the Actions tab, unless triggered to move the
built wheels to TestPyPI/PyPI.

It may also be possible to write a script that will pre-install
NonLinLoc - could be worth installing, as it would streamline the
process of installing QM almost entirely!

One step closer to pip install.