:wrench: Remove explicit PyQt5 dependency
Remove explicit PyQt5 dependency. This action is being taken to try
avoid potential clashes/crashes with pip/conda interactions.

Add a small additional printout to the `tests_import.py` script that
notifies a user if they need to install a Qt backend, and the
recommended method of doing so (`conda install pyqt`).

See comment in PR #138 for more details.