:wrench: Switch openmp -> llvm-openmp
Switch the conda install for OpenMP from openmp to llvm-openmp, which
seems to be the actively maintained version with versions that
specifically support macOS systems with M1/M2 CPUs (Arm64 architecture).

Add a MANIFEST.in file to ensure the C source code files are included in
the source distribution on PyPI, in case the wheel building fails and it
falls back on an installation from source.

Bump version to 1.0.1rc5