:bookmark_tabs: Update README/docs index
Update package README with a number of badges, quick installation
instructions, and some changes to wording to unify the language used
between the README.md file and the index page of the documentation.

Dropped the Binder sprite from the link badges, as this was just adding
a lot of clutter to the raw file.

Streamline citation section. In future we should add a CITATION.cff
file to the repo.