fix(create_lut): Add check that all stations are in grid
Add a check that all stations are contained within the grid, if "1dfmm"
function is chosen for traveltime computation. This is required, but if
not specified correctly the function otherwise didn't fail out, and
instead gave junk traveltimes (through an index overflow).

This fixes this, as outlined in https://github.com/QuakeMigrate/QuakeMigrate/issues/144