fix(export/to_mfast.py): Add units arg + minor fixes
This adds a units argument to the `sac_mfast()` function (as in `to_obspy()`
to correct station elevations to be in units of km. Also updates the reference
to the waveform file path, minor linting, plus convert a couple of header values
to strings to allow the file to be written successfully.

(This maybe should be floats instead..)