fix(plot/trigger.py): Force tz-naive timestamps for consistent type
Presumably associated with pandas version bump, any datetime string that
ends in "Z" (i.e. `'2014-08-24T00:00:00.000000Z'` or similar) is converted
to a tz-aware `pd.TimeStamp` object. When a standard datetime object is
appended to a list of these, the inconsistent tz-awareness means the final
value is ignored, meaning the availability line misses the last timestep.

This fixes this particular instance - there may be other places where this
affects appending to lists of Timestamps.