test(Volcanotectonic_Iceland): Update benchmarks
This updates the benchmark for the dike example to accommodate the changes
that were introduced by a pyproj version change, leading to some float
spillovers, which consequently caused changes in the index of the optimal
location, and then consequent changes in the locate outputs
(see https://github.com/QuakeMigrate/QuakeMigrate/issues/130).

Also update trigger benchmark - this is a pandas dtype change.