:bug: Fix bug with trace ordering
Fix a bug related to the order of traces in the streams read in during
the benchmark testing. This appears to only be an issue on macOS systems
with an M1/M2 processor. This can be fixed by explicitly calling
`.sort()` on the benchmark and test streams when they are read in.

Once added, all tests run with no issues on both macOS (12.5.1 / M1) and
Ubuntu (20.04).