:wrench:/:bookmark_tabs: Apply Black and bump copyright years
Apply the Black autoformatting tool to the entire codebase. This touches
every file, but will mean future development work is consistently styled
and removes the need for meticulous linting/style changes at the PR
level. Acceptable code line-length was set to 88 characters (up from
80), with all code comments/function docs updated to match this change.

While touching every file, it also made sense to bump the year range in
the copyright notices. In future, I think it makes sense to only worry
about bumping these years when a file is touched i.e. the years in the
range will denote the first year the work was published and the last
year in which changes were made to that particular source file.

Co-authored-by: TomWinder <tom.winder@esc.cam.ac.uk>