:test_tube: Experimental GitHub workflow for tests
Add experimental workflow for installing QuakeMigrate, running the
benchmarked examples (including an assessment of the coverage these
examples provide of the package), and run the (currently limited) tests.

While there is a current lack of unit tests, comparison of the final
outputs of the examples with the benchmarked values does give us a
holistics mean of assessing whether or not changes made have affected
the internal workings of the package.

In future we might consider expanding our unit tests, as well as
providing a slightly more streamlined means of running our tests (maybe
a command-line entry point like:

`quakemigrate run tests`