Add GitHub workflow for testing/code coverage assessment (#157)
:test_tube: Add GitHub workflow for tests/coverage

This commit implements a GitHub Action the runs the QuakeMigrate examples, assesses the code coverage these examples provide, and runs the benchmark comparison test. The Action is triggered by any of the following:

Pushes to branches prefixed with:
- feature/
- bugfix/
- patch/
- tests/

Pull requests where the target branch is either:
- master
- or is prefixed with version/
- The publication of a release

And, finally, the tests are scheduled to run once a week to keep track of breaking changes made in dependencies.

Tests are run for Python versions 3.8, 3.9, 3.10, and 3.11. Currently these tests are only run on the latest Ubuntu operating system—it would likely be wise to extend this to at least macOS, if not Windows as well (though we would have to find a solution to the NonLinLoc install issue with Windows.

In future we might consider expanding our unit tests, as well as
providing a slightly more streamlined means of running our tests (maybe
a command-line entry point like:

`quakemigrate run tests`

We might consider adjusting the installation documentation to encourage people to install NonLinLoc via the Cmake --> Make pipeline described on the package's GitHub repository.