:wrench: Add tmate to test workflow
Add tmate to test workflow. On failure, a tmate session will be
generated that can be ssh'd into to perform a debug.

This is being added to address the current issue causing our tests to
fail wherein there appears to be a slight numerical issue with the
picked times for the Iceland icequake example.