:wrench: Bypass pick file benchmark test
Bypass the pick file benchmark tests for the volcano-tectonic example,
for now, on Python 3.9 and 3.11 platforms. There appears to be some
issue with this on the GitHub workflow platform, where a single pick
differs in time in the 6th decimal place (1 microsecond). The source of
this is currently unknown, but its existence means the regularly
scheduled tests are not able to highlight other breaking changes in an
effective manner.

The picking is still being validated with the Iceland icequake
benchmarks.