:atom: Add new synthetic example
Add a new example, which uses synthetic waveforms. This example is the
one used in the manuscript to demonstrate the different stages of
QuakeMigrate. The example directory contains a small module for the
waveform simulation.

The functions in this module use type-hinting, with the docstrings
differing slightly from the standard used elsewhere in the package. A
future chore might be to type-hint and update docstrings throughout the
package. The example has been tested on Python versions 3.8 and 3.11.

This example paves the way for a comprehensive section in the online
documentation.