:bookmark_tabs: Add synthetic example documentation
Add documentation covering the synthetic example added in the previous
commit. This provides a commentary for the scripts, and should help to
expand somewhat on why certain parameter choices are made.

There is some room for improvement/extension/additions to the text, but
I think it makes a good start in the current form and should be merged
with only a brief pass to catch any critical issues (they were very much
written as stream-of-consciousness, so could benefit from a little
tightening up!).