:atom: WIP: Command-line interface
A functioning command-line interface for QuakeMigrate. This currently
offers an alternative means of initiating the various stages of a
QuakeMigrate run using `.toml` config files, instead of directly editing
`.py` scripts.

This CLI, which has two synonymous entry points (`qmctl` and
`quakemigrate`) offers:

- A means of initialising a new project using `qmctl init`. The user can
  then specify the project name, as well as various input files such as
  a station file, a velocity model file, etc, and a new directory will
  be created with example `.toml` files ready to be populated with the
  desired parameters. This directory will be marked as a QM project
  directory using a simple `.`-lock file.
- A means of then running each stage using `qmctl run`, e.g., `qmctl run
  -s detect` will seek a `detect.toml` file in the current directory and
  use the parameters defined therein to initiate a detect run.

The motivation here is to provide a user-experience that does not
require someone to be overly familiar with Python scripting. By editing
the fields of the `.toml` files in a text editor, they can get going
without having to write any code at all.

This is a work-in-progress, with a number of things still left to
resolve, including: what are sensible default `.toml` files (currently
stored in the `quakemigrate/assets` directory); documenting this way of
using QuakeMigrate.

Part of my motivation was to write the utility functions that parse the
run parameters for each stage from a `dict` style structure that could
potentially be re-used if we build a web interface for using
QuakeMigrate in the future.