:bug: Fix workflow issue with NLL compiler
Fix issue with NonLinLoc compiler in the GitHub Action workflow file. A
recent change to the NonLinLoc source code placed a hardcoded reference
to a specific GCC version. It seems the change from clang to gcc was
necessary to stop compile failures.

I tested the NonLinLoc install on my local machine first by changing the
compiler to clang, but this led to the aforementioned issues. Changing
to gcc (which I happen to have installed on my system, but would not
necessarily be a default on a macOS system and would require something
like brew to be installed.

This worked locally, but may not immediately work on the workflow
machines used by GitHub Actions. On macOS, the `sed` command had to
actually be `sed -i '' 's@/usr/local/bin/gcc-13@gcc@g' CMakeList.txt`,
but this shouldn't be the case on Ubuntu.