:atom:/:bug: Add signal transform, move onset log
Add the ability to specify the form of the signal transformation applied
to the waveform data prior to onset function calculation. These new
options include:

- energy (square of waveform amplitudes) -> this was the previous
  default
- absolute waveform amplitudes
- (complex) envelope of waveform amplitudes
- square of the complex) envelope of waveform amplitudes

Further, we changed the minimum onset value that is used as a minimum
SNR filter (previously hard-coded as 0.8), which can be set via a class
variable on the STALTAOnset class.

The 'null' value for the onset function has been changed from 0.0 to 1.0
as well.

The application of the log function to the onset functions (performed in
order to transform the product, as the geometric mean, to a summation,
as the arithmetic mean, when stacking onset functions) has been moved to
the wrapper function for the C `migrate` function. This also fixed a bug
where the log was being taken before the S onsets were combined.

In sum, these changes address a bug and introduce a number of
improvements to the flexibility of the onset function computation. They
do, however, also change the outputs of the code, and consequently the
benchmarks will need to be updated.