:wrench: Update Iceland icequake example
Update Iceland icequake example with a few parameter changes, and bring
the notebook up to date.

The only changes were made to the run scripts, no changes were made to
the core code, so all benchmarks were updated. All tests continue to
pass.

Parameter changes were made to reduce the compute time for the example
(reducing the scan timespan from 30 to 10 seconds), and adjusting the
minimum inter-event time to be shorter. This is in addition to the
earlier changes that were made to increase the resolution (and hence
refine the resultant hypocentre location!).